refactor homepage to separate client logic from metadata
